+++ client/src/Admin/AdminRosters/AdminRosterDetail.jsx
@@ -2,19 +2,24 @@ import React, { Component } from "react";
 import appClient from "../../appClient";
 import Spinner from "../../Spinner/Spinner";
 import AdminRoster from "./AdminRoster";
+import SearchTable from "../../SearchTable/SearchTable";
 import { Link } from "react-router-dom";
 import {
   UncontrolledDropdown,
   DropdownToggle,
   DropdownMenu,
-  DropdownItem
+  DropdownItem,
+  Modal,
+  ModalHeader,
+  ModalBody
 } from "reactstrap";
 import { CopyToClipboard } from "react-copy-to-clipboard";
 
 class AdminRosterDetail extends Component {
   state = {
     camp: {},
-    isLoading: false
+    isLoading: false,
+    modalOpen: false
   };
 
   componentDidMount() {
@@ -39,9 +44,18 @@ class AdminRosterDetail extends Component {
       });
   };
 
+  toggleModal = () => {
+    this.setState({
+      modalOpen: !this.state.modalOpen
+    });
+  };
+
   render() {
-    let emailAddresses = (this.state.camp.registrations || [])
+    let rosterEmails = (this.state.camp.registrations || [])
       .filter(registration => registration.paid || registration.deposit)
+      .map(registration => registration.user.email);
+    let depositEmails = (this.state.camp.registrations || [])
+      .filter(registration => !registration.paid && registration.deposit)
       .map(registration => registration.user.email);
     return (
       <div className="admin-roster-detail">
@@ -50,7 +64,21 @@ class AdminRosterDetail extends Component {
           <div className="card-header">
             <div className="d-flex justify-content-between align-items-center">
               <h5 className="card-title mb-0">Roster Details</h5>
-              <div>
+              <div className="d-flex">
+                <button
+                  className="btn btn-info mr-2"
+                  onClick={this.toggleModal}
+                >
+                  Bus
+                </button>
+                <Modal isOpen={this.state.modalOpen} toggle={this.toggleModal}>
+                  <ModalHeader toggle={this.toggleModal}>
+                    Bus Roster
+                  </ModalHeader>
+                  <ModalBody>
+                    <BusList items={this.state.camp.registrations} />
+                  </ModalBody>
+                </Modal>
                 <UncontrolledDropdown>
                   <DropdownToggle caret>Options</DropdownToggle>
                   <DropdownMenu right>
@@ -85,8 +113,11 @@ class AdminRosterDetail extends Component {
                     >
                       Download Swimming Report
                     </DropdownItem>
-                    <CopyToClipboard text={emailAddresses.join("; ")}>
-                      <DropdownItem>Copy Email Addresses</DropdownItem>
+                    <CopyToClipboard text={rosterEmails.join("; ")}>
+                      <DropdownItem>Copy Roster Emails</DropdownItem>
+                    </CopyToClipboard>
+                    <CopyToClipboard text={depositEmails.join("; ")}>
+                      <DropdownItem>Copy Deposit Emails</DropdownItem>
                     </CopyToClipboard>
                     <DropdownItem divider />
                     <Link to={`/admin/rosters/${this.state.camp._id}/swimming`}>
@@ -121,3 +152,34 @@ class AdminRosterDetail extends Component {
 }
 
 export default AdminRosterDetail;
+
+const BusList = props => {
+  const items = props.items.filter(
+    item => !item.waitlist && (item.paid || item.deposit)
+  );
+  return (
+    <SearchTable
+      items={items || []}
+      searchKeys={["camper.firstName", "camper.lastName", "afternoonPickup"]}
+      queryPlaceholder="Search List"
+      columns={[
+        {
+          key: "camper.firstName",
+          name: "First Name",
+          displayFunc: item => item.camper.firstName
+        },
+        {
+          key: "camper.lastName",
+          name: "Last Name",
+          displayFunc: item => item.camper.lastName
+        },
+        {
+          key: "afternoonPickup",
+          name: "Afternoon Pickup",
+          displayFunc: item => item.afternoonPickup
+        }
+      ]}
+      modal
+    />
+  );
+};